requirements.txt
Authlib==1.5.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
